# IGNITE-24167 Sql. Introduce heuristics to optimize join order