--- a/modules/sql-engine/src/integrationTest/java/org/apache/ignite/internal/sql/engine/statistic/ItStatisticTest.java
+++ b/modules/sql-engine/src/integrationTest/java/org/apache/ignite/internal/sql/engine/statistic/ItStatisticTest.java
@@ -49,13 +49,11 @@ public class ItStatisticTest extends BaseSqlIntegrationTest {
         long prevValueOfThreshold = sqlStatisticManager.setThresholdTimeToPostponeUpdateMs(0);
         try {
             insertAndUpdateRunQuery(500);
-            // Minimum row count is 1000, even we have less rows.
             assertQuery(getUniqueQuery())
                     .matches(scanRowCount("PUBLIC", "T", 500))
                     .check();
 
             insertAndUpdateRunQuery(600);
-            // Should return actual number of rows in the table.
             assertQuery(getUniqueQuery())
                     .matches(scanRowCount("PUBLIC", "T", 1100))
                     .check();