--- a/modules/sql-engine/src/integrationTest/sql/group1/join/test_not_distinct_from.test
+++ b/modules/sql-engine/src/integrationTest/sql/group1/join/test_not_distinct_from.test
@@ -45,7 +45,9 @@ NULL	2
 
 #Vector with vector
 query III
-select a.a, b.b, a.a IS NOT DISTINCT FROM b.b AS "Is Not Distinct From" FROM (VALUES (1), (2), (NULL)) AS a (a), (VALUES(1), (2), (NULL)) AS b (b) ORDER BY a.a NULLS LAST;
+SELECT a.a, b.b, a.a IS NOT DISTINCT FROM b.b AS "Is Not Distinct From" 
+  FROM (VALUES (1), (2), (NULL)) AS a (a), (VALUES(1), (2), (NULL)) AS b (b)
+ ORDER BY a.a NULLS LAST, b.b NULLS LAST;
 ----
 1	1	true
 1	2	false
@@ -58,7 +60,9 @@ NULL	2	false
 NULL	NULL	true
 
 query III
-select a.a, b.b, a.a IS DISTINCT FROM b.b AS "Is Distinct From" FROM (VALUES (1), (2), (NULL)) AS a (a), (VALUES (1), (2), (NULL)) AS b (b) ORDER BY a.a NULLS LAST;
+SELECT a.a, b.b, a.a IS DISTINCT FROM b.b AS "Is Distinct From" 
+  FROM (VALUES (1), (2), (NULL)) AS a (a), (VALUES (1), (2), (NULL)) AS b (b) 
+ ORDER BY a.a NULLS LAST, b.b NULLS LAST;
 ----
 1	1	false
 1	2	true