--- a/modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/metadata/IgniteMdRowCount.java
+++ b/modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/metadata/IgniteMdRowCount.java
@@ -59,8 +59,8 @@ public class IgniteMdRowCount extends RelMdRowCount {
 
     /** {@inheritDoc} */
     @Override
-    public Double getRowCount(Join rel, RelMetadataQuery mq) {
-        return rel.estimateRowCount(mq);
+    public @Nullable Double getRowCount(Join rel, RelMetadataQuery mq) {
+        return joinRowCount(mq, rel);
     }
 
     /** {@inheritDoc} */