… modules/runner/src/integrationTest/java/org/apache/ignite/internal/benchmark/AbstractTpcBenchmark.java
… modules/sql-engine/src/integrationTest/java/org/apache/ignite/internal/sql/engine/datatypes/tests/BaseIndexDataTypeTest.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/prepare/PlannerHelper.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/prepare/PlannerPhase.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/prepare/PlanningContext.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/rule/NestedLoopJoinConverterRule.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/rule/logical/IgniteMultiJoinOptimizeBushyRule.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/statistic/SqlStatisticManagerImpl.java
… modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/exec/ExecutionServiceImplTest.java
… modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/framework/TestBuilders.java
… modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/planner/AbstractPlannerTest.java
… modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/planner/AbstractTpcQueryPlannerTest.java
… modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/planner/JoinCommutePlannerTest.java
… modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/planner/TpcdsQueryPlannerTest.java
… modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/planner/TpchQueryPlannerTest.java
… modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/statistic/SqlStatisticManagerImplTest.java
… modules/sql-engine/src/test/resources/mapping/table_identity.test
… modules/sql-engine/src/test/resources/mapping/table_identity_single.test
… modules/sql-engine/src/test/resources/mapping/table_single.test
… modules/sql-engine/src/test/resources/mapping/test_backup_mapping.test
… modules/sql-engine/src/test/resources/mapping/test_partition_pruning.test
… modules/sql-engine/src/test/resources/tpcds/plan/q64.plan
… modules/sql-engine/src/test/resources/tpch/plan/q5.plan
… modules/sql-engine/src/test/resources/tpch/plan/q7.plan
… modules/sql-engine/src/test/resources/tpch/plan/q8.plan
… modules/sql-engine/src/test/resources/tpch/plan/q9.plan
… modules/sql-engine/src/testFixtures/java/org/apache/ignite/internal/sql/BaseSqlIntegrationTest.java
… modules/sql-engine/src/testFixtures/java/org/apache/ignite/internal/sql/engine/util/TpcScaleFactor.java
… modules/sql-engine/src/testFixtures/java/org/apache/ignite/internal/sql/engine/util/TpcTable.java
… modules/sql-engine/src/testFixtures/java/org/apache/ignite/internal/sql/engine/util/tpcds/TpcdsTables.java