https://issues.apache.org/jira/browse/IGNITE-24167

Benchmark results for a few queries on my laptop as follow:
```
Before
Benchmark          (fsync)  (queryId)  Mode  Cnt       Score      Error  Units
TpchBenchmark.run    false          5  avgt   20  167251.070 ± 7210.703  ms/op
TpchBenchmark.run    false          7  avgt   20   11808.586 ±  141.253  ms/op
TpchBenchmark.run    false          8  avgt   20    7457.576 ±  140.864  ms/op
TpchBenchmark.run    false          9  avgt   20   18017.645 ±  537.820  ms/op

Benchmark           (fsync)  (queryId)  Mode  Cnt      Score     Error  Units
TpcdsBenchmark.run    false         64  avgt   20  53087.853 ± 853.529  ms/op

After
Benchmark          (fsync)  (queryId)  Mode  Cnt     Score      Error  Units
TpchBenchmark.run    false          5  avgt   20  3824.434 ±   34.688  ms/op
TpchBenchmark.run    false          7  avgt   20  5771.650 ±   19.906  ms/op
TpchBenchmark.run    false          8  avgt   20  3844.181 ±   19.312  ms/op
TpchBenchmark.run    false          9  avgt   20  9358.626 ± 1251.145  ms/op

Benchmark           (fsync)  (queryId)  Mode  Cnt     Score     Error  Units
TpcdsBenchmark.run    false         64  avgt   20  4280.076 ± 112.024  ms/op
```
Note: both runs contains changes related to `SqlStatisticManagerImpl` and query texts (i.e. include hints `NO_INDEX` and `DISABLE_RULE('MergeJoinConverter')`), so only order of joins did came into play. 

---------------
Thank you for submitting the pull request.

To streamline the review process of the patch and ensure better code quality
we ask both an author and a reviewer to verify the following:

### The Review Checklist
- [ ] **Formal criteria:** TC status, codestyle, mandatory documentation. Also make sure to complete the following:  
\- There is a single JIRA ticket related to the pull request.  
\- The web-link to the pull request is attached to the JIRA ticket.  
\- The JIRA ticket has the Patch Available state.  
\- The description of the JIRA ticket explains WHAT was made, WHY and HOW.  
\- The pull request title is treated as the final commit message. The following pattern must be used: IGNITE-XXXX Change summary where XXXX - number of JIRA issue.
- [ ] **Design:** new code conforms with the design principles of the components it is added to.
- [ ] **Patch quality:** patch cannot be split into smaller pieces, its size must be reasonable.
- [ ] **Code quality:** code is clean and readable, necessary developer documentation is added if needed.
- [ ] **Tests code quality:** test set covers positive/negative scenarios, happy/edge cases. Tests are effective in terms of execution time and resources.

### Notes
- [Apache Ignite Coding Guidelines](https://cwiki.apache.org/confluence/display/IGNITE/Java+Code+Style+Guide)